Move the Docker logic into a separate module
